Add sentry recommendations given the current code
* Objective

This commit objective is about adding some recommendations on the
existing code so we could have different spans within a transaction.

* Why

We are trying to explore further the potential of Sentry's transactions
& spans, so we could use them in a real use-case.

This changes make an small move, trying to make the case about how to
create more spans in the different parts that we are meant to track.

* How

This commit makes use of the existing context within the echo request to
start new Spans in downstream layers.
