+++ middlewares/sentrytransaction.go
@@ -9,9 +9,12 @@ import (
 func SentryTransaction() echo.MiddlewareFunc {
 	return func(next echo.HandlerFunc) echo.HandlerFunc {
 		return func(ctx echo.Context) error {
-
-			span := sentry.StartSpan(
-				ctx.Request().Context(), "http", sentry.TransactionName(ctx.Request().URL.Path),
+			// This call is:
+			// - Starting a transaction.
+			// - Starting a initial spam.
+			// - Writing both in the existing echo.Context (within Request's context)
+			span := sentry.StartSpan(ctx.Request().Context(), "http",
+				sentry.TransactionName(ctx.Request().URL.Path),
 			)
 			defer span.Finish()
 			return next(ctx)
